Refactored for stats schema updates
Refactored the existing routes that serve synthetic statistics to query
using the new stats schema. In most cases this was either a change of
the table name being used (switched to using views instead) or a
simplification since fewer joins are now needed to obtain the same data
from the views.

I added one extra route to the synth API to server a list of disease
names that we track statistics for. See .. synth/diseases/list.

As a side note I also removed semicolons after the log.debug()
statements. These semicolons do not affect the execution of the code
(python just ignores them) but they are not syntactically correct. Must
have been added by a Java developer ;)

The API passes all tests in test_rest.sh and the routes return the same
data as previously expected.